Refactor form fields and buttons
