+++ src/components/index.ts
@@ -1,11 +1,12 @@
-export {default as Button} from './buttons/Button';
-export {default as ButtonRow} from './buttons/ButtonRow';
+export {default as Button} from './buttons/button';
+export {default as ButtonRow} from './buttons/button-row';
 
-export {default as Checkbox} from './formFields/check-box';
-export {default as DropDown} from './formFields/drop-down/drop-down';
-export {default as TextArea} from './formFields/TextArea';
-export {default as TextInput} from './formFields/TextInput';
-export {default as TriggerIcon} from './formFields/TriggerIcon';
+export {default as DatePicker} from './form-fields/date-picker/date-picker';
+export {default as Checkbox} from './form-fields/check-box';
+export {default as DropDown} from './form-fields/drop-down/drop-down';
+export {default as TextArea} from './form-fields/text-area';
+export {default as TextInput} from './form-fields/text-input';
+export {default as TriggerIcon} from './form-fields/trigger-icon';
 
 export {default as ButtonLink} from './links/ButtonLink';
 export {default as TextLink} from './links/TextLink';
@@ -18,6 +19,6 @@ export {default as MessageQueue} from './messages/MessageQueue';
 export {default as MobileNavbar} from './navbar/MobileNavbar';
 export {default as RegularNavbar} from './navbar/RegularNavbar';
 
-export {default as LoaderOverlay} from './overlays/LoaderOverlay';
+export {default as LoaderOverlay} from './overlays/loader-overlay';
 
 export {default as MainWrapper} from './wrapper/MainWrapper';
